feat: add use Guards
