+++ Readme.md
@@ -230,7 +230,7 @@ $ curl -s -o /dev/null -w "%{http_code}" \
 
 <br/>
 
-### Пробуем подключиться пользователем отсутствующем в базе данных fake/fake
+### Пробуем подключиться пользователем, отсутствующим в базе данных fake/fake
 
 ```
 $ curl -s -o /dev/null -w "%{http_code}"  \
@@ -249,61 +249,25 @@ $ curl -s -o /dev/null -w "%{http_code}"  \
 
 ## Комментарии к задачам:
 
-:heavy_check_mark: Пароли пользователей сохраняются в базу в виде хэша с использованием bcrypt. +20 баллов.
+:heavy_check_mark: Guards должны использоваться для работы с авторизацией/аутентификацией +30 баллов
 
-См. user.model.ts
+:heavy_check_mark: В приложении должны использоваться модули для разбиения структуры приложения на различные части (User, Board, Task) +30 баллов
 
-https://github.com/webmak1/Task-8-Authentification-JWT/blob/dev/src/resources/users/user.model.ts
+:x:Exception filters должны использоваться для обработки исключений в приложении +30 баллов
 
-```
-  @BeforeInsert()
-  async hashPassword(): Promise<void> {
-    this.password = await hash(this.password, 10);
-  }
-```
+:heavy_check_mark: @nestjs/typeorm должен использоваться для работы с базой данных +30 баллов
 
-:heavy_check_mark: Добавлен роут /login, связанная с ним логика разделена между контроллером (middleware) и соответствующим сервисом. В случае отсутствия юзера в БД, возвращается 403 (Forbidden) HTTP статус. +20 баллов.
+:x: Для логирования может использоваться встроенный Logger или кастомная имплементация. +30 баллов
 
-:heavy_check_mark: JWT токен содержит userId и login, секретный ключ хранится в .env +20 баллов.
+:heavy_check_mark: Все внешние зависимости для модулей/классов должны предоставляться с помощью механизма dependency injection. +30 баллов
 
-:heavy_check_mark: Доступ ко всем роутам, за исключением /login, /doc и /, требует аутентификации +20 баллов.
+:x:В зависисимости от env переменной USE_FASTIFY Nest.js должен использовать или express или fastify +30 баллов
 
-:heavy_check_mark: Проверка на наличие токена в реквесте реализована в отдельной middleware на уровне приложения.
-В случае если токен не валидный, или отсутствует, возвращается 401 (Unauthorized) HTTP статус. +20 баллов.
+:x:Необходимо сравнить производительность Nest.js с использованием express и fastify (можно использовать для этих целей artillery) +30 баллов
 
 <br/>
 
-### Штрафы:
-
-:heavy_check_mark: Наличие изменений в тестах либо в workflow минус 100 баллов
-
-:heavy_check_mark: Внесение изменений в репозиторий после дедлайна не считая коммиты, вносящие изменения только в Readme.md минус 30% от максимального балла за задание (для этого задания 30 баллов)
-
-:heavy_check_mark: За каждую ошибку линтера при запуске npm run lint на основе локального конфига -20 баллов (именно errors, не warnings)
-
-:heavy_check_mark: За каждую ошибку компилятора -20 баллов
-
-:heavy_check_mark: Для успешного прохождения тестов обязательно наличие в БД юзера с логином - admin, паролем - admin. Все тесты npm run test:auth должны проходить успешно, каждый не пройденный тест минус 20 баллов.
-
-Пользователь admin/admin создается при старте приложения.
-
-Выполняется скрипт.
-
-https://github.com/webmak1/Task-8-Authentification-JWT/blob/dev/src/seeds/1624496892902-CreateMigrations.ts
-
-:heavy_check_mark: Имеются явно указанные типы any, unknown -20 баллов за каждое использование
-
-:heavy_check_mark: За отсутствие отдельной ветки для разработки -20 баллов
-
-:heavy_check_mark: За отсутствие Pull Request -20 баллов
-
-:heavy_check_mark: За неполную информацию в описании Pull Request (отсутствует либо некорректен один из 3 обязательных пунктов) -10 баллов
-
-:heavy_check_mark: Меньше 3 коммитов в ветке разработки, не считая коммиты, вносящие изменения только в Readme.md — -20 баллов
-
-<br/>
-
-### Шаги
+### Шаги по созданию
 
 <br/>
