+++ app/server/.env
@@ -3,6 +3,7 @@ PORT=4000
 AUTH_MODE=true
 JWT_SECRET_KEY=JWT_SECRET_KEY
 
+USE_FASTIFY=false
 
 DATABASE_HOST=postgres
 DATABASE_NAME=db
