+++ app/server/src/boards/boards.controller.ts
@@ -8,8 +8,10 @@ import {
   Post,
   Put,
   Res,
+  UseGuards,
 } from '@nestjs/common';
 import { Response } from 'express';
+import { AuthGuard } from 'guards/auth.guard';
 import { BoardsService } from './boards.service';
 import { CreateBoardDto } from './dto/create-board.dto';
 import { UpdateBoardDto } from './dto/update-board.dto';
@@ -19,16 +21,19 @@ export class BoardsController {
   constructor(private readonly boardsService: BoardsService) {}
 
   @Get()
+  @UseGuards(AuthGuard)
   async getAllBoards() {
     return this.boardsService.getAllBoards();
   }
 
   @Post()
+  @UseGuards(AuthGuard)
   async createBoard(@Body() createBoardDto: CreateBoardDto) {
     return this.boardsService.createBoard(createBoardDto);
   }
 
   @Get(':id')
+  @UseGuards(AuthGuard)
   async getBoardById(@Res() res: Response, @Param('id') boardId: string) {
     try {
       return res
@@ -40,6 +45,7 @@ export class BoardsController {
   }
 
   @Put(':id')
+  @UseGuards(AuthGuard)
   async updateBoard(
     @Param('id') boardId: string,
     @Body() updateBoardDto: UpdateBoardDto,
@@ -48,6 +54,7 @@ export class BoardsController {
   }
 
   @Delete(':id')
+  @UseGuards(AuthGuard)
   async removeBoard(@Param('id') boardId: string) {
     return this.boardsService.removeBoard(boardId);
   }
